requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.7.4
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
daphne==4.0.0
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==5.0.2
django-allauth==0.52.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
oauthlib==3.2.2
pillow==10.2.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
zope.interface==6.4.post2
